// frontend/src/app/api/model/bookUpdateDto.ts
/**
 * Book Shelf API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BookUpdateDto { 
    title?: string;
    isbn?: string;
    pages?: number;
    yearPublished?: number;
    summary?: string;
    format?: BookUpdateDto.FormatEnum;
    authorId?: string;
    languageId?: string;
    publisherId?: string;
    genreIds?: Set<string>;
}
export namespace BookUpdateDto {
    export const FormatEnum = {
        Hardcover: 'HARDCOVER',
        Paperback: 'PAPERBACK',
        Digital: 'DIGITAL'
    } as const;
    export type FormatEnum = typeof FormatEnum[keyof typeof FormatEnum];
}